+++ src/Console/Prfs/ActionOxygenDeleteAllPrfs.php
@@ -17,8 +17,7 @@ class ActionOxygenDeleteAllPrfs extends Action {
 		if (!is_dir($f)) Oxygen::MakeLogFolder();
 		$a = glob("$f/*.prf");
 
-		foreach ($a as $ff)
-			unlink($ff);
+		if (is_array($a)) foreach ($a as $ff) unlink($ff);
 
 		Oxygen::Refresh();
